+++ chains/Cargo.toml
@@ -6,6 +6,9 @@ edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+rpc = []
+
 [dependencies]
 farcaster_core = { path = "../core" }
 strict_encoding = "1.2.1"
