+++ chains/run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+cargo test --test transactions --features rpc -- --show-output --test-threads=1
