--- chains/tests/run.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-cargo test -- --show-output --ignored --test-threads=1
